requirements.txt
starlette==0.13.6
uvicorn==0.11.8
pillow==7.2.0
python-dotenv==0.14.0
requests==2.24.0